Refactor: Implement MessageHandler class to process chat messages, convert AI message structures for various providers, transcribe voice messages, and manage message cache.
