+++ package.json
@@ -38,18 +38,17 @@
     "build": "rimraf ./build && tsc"
   },
   "dependencies": {
-    "@anthropic-ai/sdk": "^0.25.2",
-    "axios": "^1.8.4",
-    "dotenv": "^16.5.0",
+    "@anthropic-ai/sdk": "^0.56.0",
+    "axios": "^1.10.0",
+    "dotenv": "^17.2.0",
     "node-cache": "^5.1.2",
-    "openai": "4.95.0",
+    "openai": "5.9.0",
     "qrcode-terminal": "^0.12.0",
-    "whatsapp-web.js": "^1.27.0",
+    "whatsapp-web.js": "^1.31.0",
     "winston": "^3.17.0"
   },
   "devDependencies": {
-    "@types/node": "^22.14.1",
-    "nodemon": "^3.1.9",
+    "@types/node": "^24.0.13",
     "rimraf": "^6.0.1",
     "ts-node": "^10.9.2",
     "typescript": "^5.8.3"
