+++ src/roboto.ts
@@ -1,40 +1,31 @@
 import { OpenaiService } from './services/openai-service';
 import { Chat, Client, GroupChat, Message, MessageMedia, MessageTypes } from 'whatsapp-web.js';
-import {
-  bufferToStream,
-  extractAnswer,
-  getContactName,
-  getUnsupportedMessage,
-  includeName,
-  logMessage,
-  parseCommand
-} from './utils';
+import { extractAnswer, includeName, logMessage, parseCommand } from './utils';
 import logger from './logger';
 import { AIConfig, CONFIG } from './config';
-import { AIAnswer, AIContent, AiMessage, AIProvider, AIRole } from './interfaces/ai-interfaces';
+import { AIAnswer, AiMessage, AIProvider } from './interfaces/ai-interfaces';
 import { ChatCompletionMessageParam } from 'openai/resources';
 import { AnthropicService } from './services/anthropic-service';
-import { ImageBlockParam, MessageParam, TextBlock } from '@anthropic-ai/sdk/src/resources/messages';
-import NodeCache from 'node-cache';
+import { MessageParam } from '@anthropic-ai/sdk/src/resources/messages';
 import { elevenTTS } from './services/elevenlabs-service';
 import { chatConfigurationManager } from './config/chat-configurations';
 import { ResponseInput } from "openai/resources/responses/responses";
-import { AITools } from "./config/openai-functions";
+import { AITools } from "./config/functions";
+import { MessageHandler } from "./message-handler";
 
 export class Roboto {
 
-  private openAIService: OpenaiService;
+  private _openAIService: OpenaiService;
   private claudeService: AnthropicService;
-  private botConfig = CONFIG.botConfig;
+  private messageHandler: MessageHandler;
   private allowedTypes = [MessageTypes.STICKER, MessageTypes.TEXT, MessageTypes.IMAGE, MessageTypes.VOICE, MessageTypes.AUDIO];
-  private cache: NodeCache;
   private groupProcessingStatus: { [key: string]: boolean } = {};
 
   public constructor() {
 
-    this.openAIService = new OpenaiService();
+    this._openAIService = new OpenaiService();
     this.claudeService = new AnthropicService();
-    this.cache = new NodeCache();
+    this.messageHandler = new MessageHandler();
   }
 
   /**
@@ -72,9 +63,10 @@ export class Roboto {
       // Evaluates whether the message type will be processed
       if (!this.allowedTypes.includes(message.type) || (isAudioMsg && !AIConfig.SpeechConfig.enabled)) return false;
 
-      const botName = chatData.isGroup
-        ? chatConfigurationManager.getBotName(chatData.id._serialized) || this.botConfig.botName
-        : this.botConfig.botName;
+      // Bot System prompt and name
+      const chatConfiguration = chatConfigurationManager.getChatConfig(chatData.id._serialized);
+      let botName = chatConfiguration?.botName || CONFIG.botConfig.botName;
+      let systemPrompt = chatConfiguration? CONFIG.getSystemPrompt(chatConfiguration.promptInfo, botName) : CONFIG.getSystemPrompt();
 
       // Evaluates if it should respond
       const isSelfMention = message.hasQuotedMsg ? (await message.getQuotedMessage()).fromMe : false;
@@ -86,39 +78,40 @@ export class Roboto {
         await new Promise(resolve => setTimeout(resolve, 2000));
       }
       this.groupProcessingStatus[chatData.id._serialized] = true;
+      await chatData.sendStateTyping();
+
 
       // Logs the message
       logMessage(message, chatData);
 
       // Evaluates if it should go to the command flow
       if (!!command) {
-        await chatData.sendStateTyping();
         await this.commandSelect(message);
         await chatData.clearState();
         return true;
       }
 
-
       // Sends message to ChatGPT
-      chatData.sendStateTyping();
-      let chatResponseString: string = await this.processMessage(chatData);
+      let messagesList: AiMessage[] = await this.messageHandler.createMessageArray(chatData, botName);
+      let chatResponseString: string = await this.sendMessagesToAI(messagesList, systemPrompt);
       let chatResponse: AIAnswer = extractAnswer(chatResponseString, botName);
 
-      if (!chatResponse) return;
+      if (!chatResponse) return false;
 
+      // If the response includes emoji reaction, react to the message
       if(chatResponse.emojiReact)
         message.react(chatResponse.emojiReact);
 
       // Evaluate if response message must be Audio or Text
       if (chatResponse.type.toLowerCase() == 'audio' && AIConfig.SpeechConfig.enabled) {
-        return this.speak(message, chatData, chatResponse.message, 'mp3');
+        return this.returnAudioResponse(message, chatResponse.message, 'mp3');
       } else {
         return this.returnResponse(message, chatResponse.message, chatData.isGroup, client);
       }
 
     } catch (e: any) {
       logger.error(e.message);
-      return message.reply('Error 😔');
+      return false;
     } finally {
       chatData.clearState();
       this.groupProcessingStatus[chatData.id._serialized] = false;
@@ -156,116 +149,17 @@ export class Roboto {
     }
   }
 
-  /**
-   * Processes an incoming message and generates an AI response based on chat context.
-   *
-   * This function is responsible for:
-   * 1. Collecting recent chat messages to build conversation context
-   * 2. Processing multiple content types (text, images, audio)
-   * 3. Handling transcription of voice messages
-   * 4. Applying chat-specific configurations (custom prompts, bot names)
-   * 5. Selecting the appropriate AI provider and formatting messages accordingly
-   *
-   * The function limits context by time (messages older than maxHoursLimit are ignored)
-   * and resets context if a "-reset" command is found in the chat history.
-   *
-   * @param chatData - The Chat object representing the conversation
-   * @returns A promise that resolves with the AI-generated response or null if no response needed
-   */
-  private async processMessage(chatData: Chat): Promise<string> {
-
-    const actualDate = new Date();
-    const analyzeImages = !AIConfig.ChatConfig.analyzeImageDisabled;
-
-    // Bot System prompt and name
-    let systemPrompt = CONFIG.getSystemPrompt();
-    let botName = this.botConfig.botName;
-
-    const chatConfiguration = chatConfigurationManager.getChatConfig(chatData.id._serialized);
-    if (chatConfiguration) {
-      botName = chatConfiguration.botName || this.botConfig.botName;
-      systemPrompt = CONFIG.getSystemPrompt(chatConfiguration.promptInfo, botName);
-      logger.debug(`Using custom configuration for ${chatConfiguration.isGroup ? 'group' : 'chat'} "${chatConfiguration.name}" with bot name "${botName}": ${systemPrompt}`);
-    }
-
-    // Initialize an array of messages
-    const messageList: AiMessage[] = [];
-
-    // Placeholder for promises for transcriptions - Image Counting
-    let transcriptionPromises: { index: number, promise: Promise<string> }[] = [];
-    let imageCount: number = 0;
-
-    // Retrieve the last 'limit' number of messages to send them in order
-    const fetchedMessages = await chatData.fetchMessages({limit: this.botConfig.maxMsgsLimit});
-    // Check for "-reset" command in chat history to potentially restart context
-    const resetIndex = fetchedMessages.map(msg => msg.body).lastIndexOf("-reset");
-    const messagesToProcess = resetIndex >= 0 ? fetchedMessages.slice(resetIndex + 1) : fetchedMessages;
-
-    for (const msg of messagesToProcess.reverse()) {
-      try {
-        // Validate if the message was written less than 24 (or maxHoursLimit) hours ago; if older, it's not considered
-        const msgDate = new Date(msg.timestamp * 1000);
-        if ((actualDate.getTime() - msgDate.getTime()) / (1000 * 60 * 60) > this.botConfig.maxHoursLimit) break;
-
-        // Checks if a message already exists in the cache
-        const cachedMessage = this.getCachedMessage(msg);
-
-        // Check if the message includes media or if it is of another type
-        const isImage = msg.type === MessageTypes.IMAGE || msg.type === MessageTypes.STICKER;
-        const isAudio = msg.type === MessageTypes.VOICE || msg.type === MessageTypes.AUDIO;
-        const isOther = !isImage && !isAudio && msg.type != 'chat';
-
-        // Limit the number of processed images to only the last few and ignore audio if cached
-        const media = (isImage && imageCount < this.botConfig.maxImages && analyzeImages) || (isAudio && !cachedMessage) ?
-          await msg.downloadMedia() : null;
-
-        if (media && isImage) imageCount++;
-
-        const role = (!msg.fromMe || isImage) ? AIRole.USER : AIRole.ASSISTANT;
-        const name = msg.fromMe ? botName : (await getContactName(msg));
-
-        // Assemble the content as a mix of text and any included media
-        const content: Array<AIContent> = [];
-        if (isOther || (isAudio && !AIConfig.TranscriptionConfig.enabled))
-          content.push({type: 'text', value: getUnsupportedMessage(msg.type, msg.body)});
-        else if (isAudio && media && !cachedMessage) {
-          transcriptionPromises.push({index: messageList.length, promise: this.transcribeVoice(media, msg)});
-          content.push({type: 'audio', value: '<Transcribing voice message...>'});
-        }
-        if (isAudio && cachedMessage) content.push({type: 'audio', value: cachedMessage});
-        if (isImage && media) content.push({type: 'image', value: media.data, media_type: media.mimetype});
-        if (isImage && !media) content.push({type: 'text', value: '<Unprocessed image>'});
-        if (msg.body && !isOther) content.push({type: 'text', value: msg.body});
-
-        messageList.push({role: role, name: name, content: content});
-      } catch (e: any) {
-        logger.error(`Error reading message - msg.type:${msg.type}; msg.body:${msg.body}. Error:${e.message}`);
-      }
-    }
-
-    // If no new messages are present, return without action
-    if (messageList.length == 0) return;
-
-    // Wait for all transcriptions to complete
-    const transcriptions = await Promise.all(transcriptionPromises.map(t => t.promise));
-    transcriptionPromises.forEach((transcriptionPromise, idx) => {
-      const transcription = transcriptions[idx];
-      const messageIdx = transcriptionPromise.index;
-      messageList[messageIdx].content = messageList[messageIdx].content.map(c =>
-        c.type === 'audio' && c.value === '<Transcribing voice message...>' ? {type: 'audio', value: transcription} : c
-      );
-    });
-
+  private async sendMessagesToAI(messageList: AiMessage[], systemPrompt): Promise<string> {
     // Send the message and return the text response
     if (AIConfig.ChatConfig.provider == AIProvider.CLAUDE) {
-      const convertedMessageList: MessageParam[] = this.convertIaMessagesLang(messageList.reverse(), AIProvider.CLAUDE) as MessageParam[];
+      const convertedMessageList: MessageParam[] = this.messageHandler.convertIaMessagesLang(messageList.reverse(), AIProvider.CLAUDE) as MessageParam[];
       return await this.claudeService.sendChat(convertedMessageList, CONFIG.getSystemPrompt());
     } else if (AIConfig.ChatConfig.provider == AIProvider.OPENAI) {
-      const convertedMessageList: ResponseInput = this.convertIaMessagesLang(messageList.reverse(), AIConfig.ChatConfig.provider as AIProvider, systemPrompt) as ResponseInput;
-      return await this.openAIService.sendChatWithTools(convertedMessageList, 'text', AITools);
+      const convertedMessageList: ResponseInput = this.messageHandler.convertIaMessagesLang(messageList.reverse(), AIConfig.ChatConfig.provider as AIProvider, systemPrompt) as ResponseInput;
+      return await this._openAIService.sendChatWithTools(convertedMessageList, 'text', AITools);
     } else {
-      const convertedMessageList: ChatCompletionMessageParam[] = this.convertIaMessagesLang(messageList.reverse(), AIConfig.ChatConfig.provider as AIProvider, systemPrompt) as ChatCompletionMessageParam[];
-      return await this.openAIService.sendCompletion(convertedMessageList);
+      const convertedMessageList: ChatCompletionMessageParam[] = this.messageHandler.convertIaMessagesLang(messageList.reverse(), AIConfig.ChatConfig.provider as AIProvider, systemPrompt) as ChatCompletionMessageParam[];
+      return await this._openAIService.sendCompletion(convertedMessageList);
     }
   }
 
@@ -280,22 +174,18 @@ export class Roboto {
    * as input for speech synthesis. The resulting audio is sent as a voice message.
    *
    * @param message - The Message object for reply context
-   * @param chatData - The Chat object for the conversation
-   * @param content - Optional text content to convert to speech
+   * @param messageToSay - Text to synthesize
    * @param responseFormat - Optional format for the audio response
    * @returns A promise that resolves when the audio message is sent
    */
-  private async speak(message: Message, chatData: Chat, content: string | undefined, responseFormat?) {
-    // Set the content to be spoken. If no content is explicitly provided, fetch the last bot reply for use.
-    let messageToSay = content || await this.getLastBotMessage(chatData);
-    try {
-      // Generate speech audio from the given text content using the OpenAI API.
+  private async returnAudioResponse(message: Message, messageToSay: string | undefined, responseFormat?): Promise<void> {
 
+    try {
       let base64Audio;
       if (AIConfig.SpeechConfig.provider == AIProvider.ELEVENLABS) {
         base64Audio = await elevenTTS(messageToSay);
       } else {
-        const audioBuffer = await this.openAIService.speech(messageToSay, responseFormat);
+        const audioBuffer = await this._openAIService.speech(messageToSay, responseFormat);
         base64Audio = audioBuffer.toString('base64');
       }
 
@@ -304,7 +194,9 @@ export class Roboto {
       // Reply to the message with the synthesized speech audio.
       const repliedMsg = await message.reply(audioMedia, undefined, {sendAudioAsVoice: true});
 
-      this.cache.set(repliedMsg.id._serialized, messageToSay, CONFIG.botConfig.nodeCacheTime);
+      // Store the message in the cache for future use.
+      this.messageHandler.cache.set(repliedMsg.id._serialized, messageToSay, CONFIG.botConfig.nodeCacheTime);
+
     } catch (e: any) {
       logger.error(`Error in speak function: ${e.message}`);
       throw e;
@@ -328,7 +220,7 @@ export class Roboto {
 
     try {
       // Calls the ChatGPT service to generate an image based on the provided textual content.
-      const imgResponse = await this.openAIService.createImage(content);
+      const imgResponse = await this._openAIService.createImage(content);
       let media;
       if (imgResponse[0].url) {
         media = await MessageMedia.fromUrl(imgResponse[0].url);
@@ -342,174 +234,6 @@ export class Roboto {
       logger.error(`Error in createImage function: ${e.message}`);
       // In case of an error during image generation or sending the image, inform the user.
       return message.reply("I encountered a problem while trying to generate an image, please try again.");
-    }
-  }
-
-  /**
-   * Retrieves the last message sent by the bot in a chat.
-   *
-   * This function fetches recent messages from the chat history and
-   * finds the most recent message sent by the bot (fromMe = true)
-   * that contains substantive content.
-   *
-   * @param chatData - The Chat object to search for messages
-   * @returns A promise resolving to the text of the last bot message
-   */
-  private async getLastBotMessage(chatData: Chat) {
-    const lastMessages = await chatData.fetchMessages({limit: 12});
-    let lastMessageBot: string = '';
-    for (const msg of lastMessages) {
-      if (msg.fromMe && msg.body.length > 1) lastMessageBot = msg.body;
-    }
-    return lastMessageBot;
-  }
-
-  /**
-   * Converts AI message structures between different language model formats.
-   *
-   * This function transforms message arrays into formats compatible with various AI providers:
-   * - OpenAI: Structured with system, user and assistant roles
-   * - Claude: Requires alternating user/assistant roles with specific content formats
-   * - Qwen: Similar to OpenAI but with provider-specific adaptations
-   * - DeepSeek: Uses JSON formatting with text blocks
-   * - DeepInfra/Custom: Uses simplified message formatting
-   *
-   * The function handles text, audio transcriptions, and images appropriately for each provider.
-   *
-   * @param messageList - Array of AI messages to convert
-   * @param lang - The target AI provider format
-   * @param systemPrompt - Optional system prompt to include
-   * @returns Formatted message array compatible with the specified AI provider
-   */
-  private convertIaMessagesLang(messageList: AiMessage[], lang: AIProvider, systemPrompt?: string): MessageParam[] | ChatCompletionMessageParam[] | ResponseInput {
-    switch (lang) {
-      case AIProvider.CLAUDE:
-
-        const claudeMessageList: MessageParam[] = [];
-        let currentRole: AIRole = AIRole.USER;
-        let gptContent: Array<TextBlock | ImageBlockParam> = [];
-        messageList.forEach((msg, index) => {
-          const role = msg.role === AIRole.ASSISTANT && msg.content.find(c => c.type === 'image') ? AIRole.USER : msg.role;
-          if (role !== currentRole) { // Change role or if it's the last message
-            if (gptContent.length > 0) {
-              claudeMessageList.push({role: currentRole as any, content: gptContent});
-              gptContent = []; // Reset for the next block of messages
-            }
-            currentRole = role; // Ensure role alternation 
-          }
-
-          // Add content to the current block
-          msg.content.forEach(c => {
-            if (['text', 'audio'].includes(c.type)) gptContent.push({
-              type: 'text',
-              text: JSON.stringify({message: c.value, author: msg.name, type: c.type})
-            });
-            if (['image'].includes(c.type)) gptContent.push({
-              type: 'image',
-              source: {data: c.value!, media_type: c.media_type as any, type: 'base64'}
-            });
-          });
-        });
-        // Ensure the last block is not left out
-        if (gptContent.length > 0) claudeMessageList.push({role: currentRole, content: gptContent});
-
-        // Ensure the first message is always AiRole.USER (by API requirement)
-        if (claudeMessageList.length > 0 && claudeMessageList[0].role !== AIRole.USER) {
-          claudeMessageList.shift(); // Remove the first element if it's not USER
-        }
-
-        return claudeMessageList;
-
-      case AIProvider.DEEPSEEK:
-
-        const deepSeekMsgList: any[] = [];
-        messageList.forEach(msg => {
-          if (msg.role == AIRole.ASSISTANT) {
-            const textContent = msg.content.find(c => ['text', 'audio'].includes(c.type))!;
-            const content = JSON.stringify({
-              type: 'text',
-              text: JSON.stringify({message: textContent.value, author: msg.name, type: textContent.type, response_format: "json_object"})
-            });
-            deepSeekMsgList.push({content: content, name: msg.name!, role: msg.role});
-          } else {
-            const gptContent: Array<any> = [];
-            msg.content.forEach(c => {
-              if (['image'].includes(c.type)) gptContent.push({
-                type: 'text',
-                text: JSON.stringify({message: getUnsupportedMessage('image', ''), author: msg.name, type: c.type, response_format: "json_object"})
-              });
-              if (['text', 'audio'].includes(c.type)) gptContent.push({
-                type: 'text',
-                text: JSON.stringify({message: c.value, author: msg.name, type: c.type, response_format: "json_object"})
-              });
-            })
-            deepSeekMsgList.push({content: gptContent, name: msg.name!, role: msg.role});
-          }
-        })
-
-        deepSeekMsgList.unshift({role: AIRole.SYSTEM, content: [{type: 'text', text: systemPrompt}]});
-
-        return deepSeekMsgList;
-
-      case AIProvider.OPENAI:
-
-        const responseAPIMessageList: ResponseInput = [];
-        messageList.forEach(msg => {
-          const gptContent: Array<any> = [];
-          msg.content.forEach(c => {
-            const fromBot = msg.role == AIRole.ASSISTANT;
-            if (['text', 'audio'].includes(c.type))  gptContent.push({ type: fromBot?'output_text':'input_text', text: JSON.stringify({message: c.value, author: msg.name, type: c.type, response_format:'json_object'}) });
-            if (['image'].includes(c.type))          gptContent.push({ type: 'input_image', image_url: `data:${c.media_type};base64,${c.value}`});
-          })
-          responseAPIMessageList.push({content: gptContent, role: msg.role});
-        })
-
-        responseAPIMessageList.unshift({role: AIRole.SYSTEM, content: systemPrompt});
-        return responseAPIMessageList;
-
-      case AIProvider.QWEN:
-        const chatgptMessageList: any[] = [];
-        messageList.forEach(msg => {
-          const gptContent: Array<any> = [];
-          msg.content.forEach(c => {
-            if (['text', 'audio'].includes(c.type)) gptContent.push({
-              type: 'text',
-              text: JSON.stringify({message: c.value, author: msg.name, type: c.type, response_format: 'json_object'})
-            });
-            if (['image'].includes(c.type)) gptContent.push({type: 'image_url', image_url: {url: `data:${c.media_type};base64,${c.value}`}});
-          })
-          chatgptMessageList.push({content: gptContent, name: msg.name!, role: msg.role});
-        })
-
-        chatgptMessageList.unshift({role: AIRole.SYSTEM, content: [{type: 'text', text: systemPrompt}]});
-
-        return chatgptMessageList;
-
-      case AIProvider.CUSTOM:
-      case AIProvider.DEEPINFRA:
-
-        const otherMsgList: any[] = [];
-        messageList.forEach(msg => {
-          if (msg.role == AIRole.ASSISTANT) {
-            const textContent = msg.content.find(c => ['text', 'audio'].includes(c.type))!;
-            const content = JSON.stringify({message: textContent.value, author: msg.name, type: textContent.type, response_format: "json_object"});
-            otherMsgList.push({content: content, name: msg.name!, role: msg.role});
-          } else {
-            const gptContent: Array<any> = [];
-            msg.content.forEach(c => {
-              if (['image'].includes(c.type)) gptContent.push(JSON.stringify({message: getUnsupportedMessage('image', ''), author: msg.name, type: c.type, response_format: "json_object"}));
-              if (['text', 'audio'].includes(c.type)) gptContent.push(JSON.stringify({message: c.value, author: msg.name, type: c.type, response_format: "json_object"}));
-            })
-            otherMsgList.push({content: gptContent[0], role: msg.role});
-          }
-        })
-
-        otherMsgList.unshift({role: AIRole.SYSTEM, content: systemPrompt});
-
-        return otherMsgList;
-
-      default:
-        return [];
     }
   }
 
@@ -532,7 +256,6 @@ export class Roboto {
     const chat = await message.getChat();
     const isGroup = chat.isGroup;
 
-    // Solo verificar permisos de administrador en grupos
     if (isGroup) {
       const groupChat = chat as GroupChat;
       const participant = await groupChat.participants.find(p => p.id._serialized === message.author);
@@ -602,55 +325,6 @@ export class Roboto {
     }
   }
 
-
-  /**
-   * Transcribes a voice message to text using AI services.
-   *
-   * This function:
-   * 1. Checks if the transcription already exists in cache
-   * 2. Converts the media to an audio buffer and stream
-   * 3. Sends the audio to the configured transcription service
-   * 4. Caches the result for future use
-   *
-   * The function uses the OpenAI service which may route to different
-   * providers based on configuration.
-   *
-   * @param media - The MessageMedia object containing the voice data
-   * @param message - The Message object for cache identification
-   * @returns A promise resolving to the transcribed text
-   */
-  private async transcribeVoice(media: MessageMedia, message: Message): Promise<string> {
-    try {
-
-      // Check if the transcription exists in the cache
-      const cachedMessage = this.getCachedMessage(message);
-      if (cachedMessage) return cachedMessage;
-
-      // Convert the base64 media data to a Buffer
-      const audioBuffer = Buffer.from(media.data, 'base64');
-
-      // Convert the buffer to a stream
-      const audioStream = bufferToStream(audioBuffer);
-
-      logger.debug(`[${AIConfig.TranscriptionConfig.provider}->transcribeVoice] Starting audio transcription`);
-
-      const transcribedText = await this.openAIService.transcription(audioStream);
-
-      // Log the transcribed text
-      logger.debug(`[${AIConfig.TranscriptionConfig.provider}->transcribeVoice] Transcribed text: ${transcribedText}`);
-
-      // Store in cache
-      this.cache.set(message.id._serialized, transcribedText, CONFIG.botConfig.nodeCacheTime);
-
-      return transcribedText;
-
-    } catch (error: any) {
-      // Error handling
-      logger.error(`Error transcribing voice message: ${error.message}`);
-      return '<Error transcribing voice message>';
-    }
-  }
-
   /**
    * Sends a response message appropriately based on chat context.
    *
@@ -669,17 +343,8 @@ export class Roboto {
     else return client.sendMessage(message.from, responseMsg);
   }
 
-  /**
-   * Retrieves a cached message by its unique identifier.
-   *
-   * This function checks the NodeCache instance for previously stored
-   * message content, such as transcriptions or generated responses.
-   *
-   * @param msg - The Message object whose content might be cached
-   * @returns The cached string content or undefined if not found
-   */
-  private getCachedMessage(msg: Message) {
-    return this.cache.get<string>(msg.id._serialized);
+  get openAIService(): OpenaiService {
+    return this._openAIService;
   }
 
 }
